+++ tgbotapi.extensions.utils/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/utils/extensions/FlowsUpdatesFilter.kt
@@ -6,14 +6,16 @@ import dev.inmo.tgbotapi.updateshandlers.FlowsUpdatesFilter
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.merge
 
+@Deprecated("Will be removed soon", ReplaceWith("messageFlow + channelPostFlow"))
 val FlowsUpdatesFilter.allSentMessagesFlow: Flow<BaseSentMessageUpdate>
     get() = merge(
-        messageFlow,
-        channelPostFlow
+        messagesFlow,
+        channelPostsFlow
     )
 
+@Deprecated("Will be removed soon", ReplaceWith("messageMediaGroupFlow + channelPostMediaGroupFlow"))
 val FlowsUpdatesFilter.allSentMediaGroupsFlow: Flow<SentMediaGroupUpdate>
     get() = merge(
-        messageMediaGroupFlow,
-        channelPostMediaGroupFlow
+        messageMediaGroupsFlow,
+        channelPostMediaGroupsFlow
     )